+++ compiler/GHC/Builtin/Names.hs
@@ -384,8 +384,6 @@ basicKnownKeyNames
         integerModName,
         integerDivModName,
         integerQuotRemName,
-        integerToFloatName,
-        integerToDoubleName,
         integerEncodeFloatName,
         integerEncodeDoubleName,
         integerGcdName,
@@ -438,6 +436,8 @@ basicKnownKeyNames
         bignatFromWordListName,
 
         -- Float/Double
+        integerToFloatName,
+        integerToDoubleName,
         rationalToFloatName,
         rationalToDoubleName,
 
@@ -1197,8 +1197,6 @@ integerFromNaturalName
    , integerModName
    , integerDivModName
    , integerQuotRemName
-   , integerToFloatName
-   , integerToDoubleName
    , integerEncodeFloatName
    , integerEncodeDoubleName
    , integerGcdName
@@ -1324,8 +1322,6 @@ integerDivName            = bniVarQual "integerDiv"                integerDivIdK
 integerModName            = bniVarQual "integerMod"                integerModIdKey
 integerDivModName         = bniVarQual "integerDivMod#"            integerDivModIdKey
 integerQuotRemName        = bniVarQual "integerQuotRem#"           integerQuotRemIdKey
-integerToFloatName        = bniVarQual "integerToFloat#"           integerToFloatIdKey
-integerToDoubleName       = bniVarQual "integerToDouble#"          integerToDoubleIdKey
 integerEncodeFloatName    = bniVarQual "integerEncodeFloat#"       integerEncodeFloatIdKey
 integerEncodeDoubleName   = bniVarQual "integerEncodeDouble#"      integerEncodeDoubleIdKey
 integerGcdName            = bniVarQual "integerGcd"                integerGcdIdKey
@@ -1370,7 +1366,9 @@ floatingClassName  = clsQual gHC_FLOAT (fsLit "Floating")  floatingClassKey
 realFloatClassName = clsQual gHC_FLOAT (fsLit "RealFloat") realFloatClassKey
 
 -- other GHC.Float functions
-rationalToFloatName, rationalToDoubleName :: Name
+integerToFloatName, integerToDoubleName, rationalToFloatName, rationalToDoubleName :: Name
+integerToFloatName   = varQual gHC_FLOAT (fsLit "integerToFloat#") integerToFloatIdKey
+integerToDoubleName  = varQual gHC_FLOAT (fsLit "integerToDouble#") integerToDoubleIdKey
 rationalToFloatName  = varQual gHC_FLOAT (fsLit "rationalToFloat") rationalToFloatIdKey
 rationalToDoubleName = varQual gHC_FLOAT (fsLit "rationalToDouble") rationalToDoubleIdKey
 
@@ -2387,6 +2385,10 @@ coercionTokenIdKey      = mkPreludeMiscIdUnique 124
 noinlineIdKey           = mkPreludeMiscIdUnique 125
 considerAccessibleIdKey = mkPreludeMiscIdUnique 126
 
+integerToFloatIdKey, integerToDoubleIdKey :: Unique
+integerToFloatIdKey    = mkPreludeMiscIdUnique 128
+integerToDoubleIdKey   = mkPreludeMiscIdUnique 129
+
 rationalToFloatIdKey, rationalToDoubleIdKey :: Unique
 rationalToFloatIdKey   = mkPreludeMiscIdUnique 130
 rationalToDoubleIdKey  = mkPreludeMiscIdUnique 131
@@ -2600,8 +2602,6 @@ integerFromNaturalIdKey
    , integerModIdKey
    , integerDivModIdKey
    , integerQuotRemIdKey
-   , integerToFloatIdKey
-   , integerToDoubleIdKey
    , integerEncodeFloatIdKey
    , integerEncodeDoubleIdKey
    , integerGcdIdKey
@@ -2683,8 +2683,6 @@ integerDivIdKey            = mkPreludeMiscIdUnique 624
 integerModIdKey            = mkPreludeMiscIdUnique 625
 integerDivModIdKey         = mkPreludeMiscIdUnique 626
 integerQuotRemIdKey        = mkPreludeMiscIdUnique 627
-integerToFloatIdKey        = mkPreludeMiscIdUnique 628
-integerToDoubleIdKey       = mkPreludeMiscIdUnique 629
 integerEncodeFloatIdKey    = mkPreludeMiscIdUnique 630
 integerEncodeDoubleIdKey   = mkPreludeMiscIdUnique 631
 integerGcdIdKey            = mkPreludeMiscIdUnique 632